Cleaning up 'optionally-blockable'.
bz@ noted in https://github.com/w3c/webappsec-mixed-content/issues/8 that the
definition of optionally-blockable was confusing, especially with regard to
the distinction between '<img>', '<img srcset>', and '<picture>'. This patch
attempts to clarify, both by explicitly referencing the Fetch properties that
are used in the algorithms below, and by adding explanatory text around the
image-related carveouts.