Update title and prose to match CRD living-standard nature
With the switch autopublishing of the current spec as a CRD at
https://www.w3.org/TR/mixed-content/, we have replaced the “Level 1”
spec in-place and effectively — and retroactively — made Mixed Content a
Living Standard.

As a consequence of that, we need to drop the “Level 2” part from the
spec, and drop all references to the TR doc as being the “Level 1” spec
that is obsoleted. Instead we need to adjust the spec prose to
refer the current spec as simply an incremental update to itself.